Gosh this is a big one.
AllMap supports resolving refs on each map.
As such ResolveRef happens only when Binding to a Map via BindToMap
BindToMap will be called every remake, and if it's the same map, is skipped.
When loaded, refs are not resolved - only the moment a search is performed . . . which is probably immediately anyway right?

Gather search params into a struct for easier passing around.
That struct holds the BaseList, map search type (Cur/All/Chosen), and if needed, a list of chosen maps

Similarly there's a struct for the results:
which holds list of maps and the things found

FindDesc editor subclassable for the header
Base editor now not setting the map settings in the editor as it's a non-active filter editor

MapLabel is just generated when displayed.