OH man oh geez
Changing the name of all the classes. Because "filter" was thrown around too much, it's now either "query" or "search"

ListFilter => ThingQuery, displayed as "a filter", asking a single thing yes/no
FindDescription => QuerySearch, "a search", a group of queries asking a list of things yes/no
FilterGroup => SearchGroup, "a group", a grouping of searches, just for organization. Can copy groups
And above that would be the "search library", the group of groupings of searches

The "Filter" string is still displayed to the user. But that's for Individual ThingQueries. "Search" is used when it's a list of queries, aka a QuerySearch.

So I think I got places where "filter" meant "query" or "search" but it's easily imaginable there's some still there saying the wrong thing,

but it works.