health condition values are usually percentages
Trying to set up a condition on hypothermia being 0.2-0.3 is strange
and also lacks precision. I don't know what this was this way,
I couldn't find a condition that'd use such values (e.g. gunshot
could be something like 1.2, but the UI condition for it doesn't even
allow a range).